Add files via upload
We will guide you through the art of harnessing OpenAI Whisper's speech recognition prowess. Together, we'll delve into the mysterious world of Python code to masterfully transcribe the words of Fed Chairman Powell, spoken on the fateful day of November 2nd. Join me on this exhilarating journey and discover the secrets of Whisper.